move pdfviewer ui data source to separate file
