+++ atom/browser/ui/webui/pdf_viewer_ui.h
@@ -1,0 +1,36 @@
+// Copyright (c) 2017 GitHub, Inc.
+// Use of this source code is governed by the MIT license that can be
+// found in the LICENSE file.
+
+#ifndef ATOM_BROWSER_UI_WEBUI_PDF_VIEWER_UI_H_
+#define ATOM_BROWSER_UI_WEBUI_PDF_VIEWER_UI_H_
+
+#include <string>
+
+#include "base/macros.h"
+#include "content/public/browser/web_ui_controller.h"
+
+namespace content {
+class BrowserContext;
+}
+
+namespace atom {
+
+class PdfViewerUI : public content::WebUIController {
+ public:
+  static const char kHost[];
+
+  PdfViewerUI(content::BrowserContext* browser_context,
+              content::WebUI* web_ui,
+              const std::string& view_id);
+
+  // content::WebUIController implementation.
+  void RenderViewCreated(content::RenderViewHost* rvh) override;
+
+ private:
+  DISALLOW_COPY_AND_ASSIGN(PdfViewerUI);
+};
+
+}  // namespace atom
+
+#endif  // ATOM_BROWSER_UI_WEBUI_PDF_VIEWER_UI_H_
